Add coverage, move database connection support.
