+++ spec/support/active_record.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
-require 'config/database_connection'
+require_relative 'database_connection'
 DatabaseConnection.connect!(ENV['RACK_ENV'])
